PN-252: Run local matches search only for patients updated/created since last search
Simplified the API a bit
