+++ matching-notification-api/src/main/java/org/phenotips/matchingnotification/finder/internal/DefaultMatchFinderManager.java
@@ -18,12 +18,16 @@
 package org.phenotips.matchingnotification.finder.internal;
 
 import org.phenotips.data.Patient;
+import org.phenotips.data.PatientRepository;
+import org.phenotips.data.permissions.PermissionsManager;
 import org.phenotips.data.permissions.Visibility;
 import org.phenotips.matchingnotification.finder.MatchFinder;
 import org.phenotips.matchingnotification.finder.MatchFinderManager;
 import org.phenotips.matchingnotification.match.PatientMatch;
 
 import org.xwiki.component.annotation.Component;
+import org.xwiki.query.Query;
+import org.xwiki.query.QueryManager;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -52,17 +56,47 @@ public class DefaultMatchFinderManager implements MatchFinderManager
     @Named("matchable")
     private Visibility matchableVisibility;
 
+    @Inject
+    private QueryManager qm;
+
+    @Inject
+    private PatientRepository patientRepository;
+
+    @Inject
+    private PermissionsManager permissionsManager;
+
+
+    @Override
+    public void findMatchesForAllPatients()
+    {
+        this.recordFindAllStart();
+
+        List<Patient> patients = this.getPatientsList();
+
+        for (Patient patient : patients) {
+            this.logger.debug("Finding matches for patient {}.", patient.getId());
+            this.findMatches(patient);
+        }
+
+        this.recordFindAllEnd();
+    }
+
+
     @Override
     public List<PatientMatch> findMatches(Patient patient)
     {
         List<PatientMatch> matches = new LinkedList<>();
 
         for (MatchFinder service : this.matchFinderProvider.get()) {
+
+            this.logger.error("Using service {}", service.getClass().getSimpleName());
+
             try {
                 List<PatientMatch> foundMatches = service.findMatches(patient);
                 matches.addAll(foundMatches);
 
-                this.logger.debug("Found {} matches by {}: ", foundMatches.size(), service.getClass().getSimpleName());
+                this.logger.error("Found {} matches by {}", foundMatches.size(), service.getClass().getSimpleName());
+
                 for (PatientMatch match : foundMatches) {
                     this.logger.debug(match.toString());
                 }
@@ -76,19 +110,51 @@ public class DefaultMatchFinderManager implements MatchFinderManager
         return matches;
     }
 
-    @Override
-    public void recordMatcherStart()
+    private void recordFindAllStart()
     {
         for (MatchFinder service : this.matchFinderProvider.get()) {
             service.recordStartMatchesSearch();
         }
     }
 
-    @Override
-    public void recordMatcherEnd()
+    private void recordFindAllEnd()
     {
         for (MatchFinder service : this.matchFinderProvider.get()) {
             service.recordEndMatchesSearch();
         }
+    }
+
+    /**
+     * Returns a list of patients with visibility >= matchable.
+     */
+    private List<Patient> getPatientsList()
+    {
+        List<Patient> patients = new LinkedList<>();
+
+        try {
+            Query q = this.qm.createQuery(
+                "select doc.name "
+                    + "from Document doc, "
+                    + "doc.object(PhenoTips.PatientClass) as patient "
+                    + "where patient.identifier is not null order by patient.identifier desc",
+                Query.XWQL);
+            List<String> potentialPatientIds = q.execute();
+
+            for (String patientId : potentialPatientIds) {
+                Patient patient = this.patientRepository.get(patientId);
+                if (patient == null) {
+                    continue;
+                }
+
+                Visibility patientVisibility = this.permissionsManager.getPatientAccess(patient).getVisibility();
+                if (patientVisibility.compareTo(this.matchableVisibility) >= 0) {
+                    patients.add(patient);
+                }
+            }
+        } catch (Exception e) {
+            this.logger.error("Error retrieving a list of patients for matching: {}", e);
+        }
+
+        return patients;
     }
 }
